api/pyproject.toml
[project]
name = "aedmi-api"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "sqlalchemy>=2.0",
    "asyncpg>=0.29",
    "pyjwt>=2.8",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pandas>=2.0",
    "openpyxl>=3.1",
    "python-multipart>=0.0.9",
    "pypdfium2>=4.30",
    "pillow>=10.0",
]

[project.optional-dependencies]
dev = [
    "hypothesis>=6.100",
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "W", "F", "I"]

[tool.ruff.lint.isort]
known-first-party = ["core", "routers", "schemas"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
